feat: migrate inventory quantity field to Decimal type
- Change inventory.quantity from Integer to Numeric(10,3) to support decimal values
- Update all schemas (InventoryBase, InventoryUpdate, InventoryRead, InventoryReadSimple, InventoryReadDetailed) to use Decimal
- Update CRUD search filters (min_quantity, max_quantity) to accept Decimal
- Update router query parameters to accept Decimal types
- Update seed data to use Decimal('0.000') for inventory quantities
- Add migration script for converting existing database
- Maintains backward compatibility with integer values

+++ app/crud/inventory.py
@@ -1,6 +1,7 @@
 from typing import List, Optional
 from sqlalchemy.orm import Session
 from sqlalchemy import select, and_, or_, func
+from decimal import Decimal
 from app.models.inventory import Inventory
 from app.schemas.inventory import InventoryCreate, InventoryUpdate
 
@@ -50,8 +51,8 @@ def search(
     q: Optional[str] = None,
     category_id: Optional[int] = None,
     weight_id: Optional[int] = None,
-    min_quantity: Optional[int] = None,
-    max_quantity: Optional[int] = None,
+    min_quantity: Optional[Decimal] = None,
+    max_quantity: Optional[Decimal] = None,
     skip: int = 0,
     limit: int = 100,
 ) -> List[Inventory]:
