+++ app/schemas/inventory.py
@@ -1,5 +1,6 @@
 from pydantic import BaseModel
 from typing import Optional, TYPE_CHECKING
+from decimal import Decimal
 
 if TYPE_CHECKING:
     from app.schemas.category import CategoryReadSimple
@@ -8,7 +9,7 @@ if TYPE_CHECKING:
 class InventoryBase(BaseModel):
     name: str
     shortname: Optional[str] = None
-    quantity: int
+    quantity: Decimal
     category_id: int
     weight_id: int
 
@@ -18,7 +19,7 @@ class InventoryCreate(InventoryBase):
 class InventoryUpdate(BaseModel):
     name: Optional[str] = None
     shortname: Optional[str] = None
-    quantity: Optional[int] = None
+    quantity: Optional[Decimal] = None
     category_id: Optional[int] = None
     weight_id: Optional[int] = None
 
@@ -26,7 +27,7 @@ class InventoryRead(BaseModel):
     id: int
     name: str
     shortname: Optional[str]
-    quantity: int
+    quantity: Decimal
     category_id: int
     weight_id: int
     model_config = {"from_attributes": True}
@@ -35,7 +36,7 @@ class InventoryReadSimple(BaseModel):
     id: int
     name: str
     shortname: Optional[str]
-    quantity: int
+    quantity: Decimal
     category_id: int
     weight_id: int
     model_config = {"from_attributes": True}
@@ -45,7 +46,7 @@ class InventoryReadDetailed(BaseModel):
     id: int
     name: str
     shortname: Optional[str]
-    quantity: int
+    quantity: Decimal
     category_id: int
     weight_id: int
     category: "CategoryReadSimple"
